fix cycle dep from NewClusterList (#2225)
