+++ module/epochs/qc_voter_test.go
@@ -13,6 +13,7 @@ import (
 
 	hotstuff "github.com/onflow/flow-go/consensus/hotstuff/mocks"
 	"github.com/onflow/flow-go/model/flow"
+	"github.com/onflow/flow-go/model/flow/factory"
 	flowmodule "github.com/onflow/flow-go/module"
 	"github.com/onflow/flow-go/module/epochs"
 	module "github.com/onflow/flow-go/module/mock"
@@ -76,7 +77,7 @@ func (suite *Suite) SetupTest() {
 
 	var err error
 	assignments := unittest.ClusterAssignment(2, suite.nodes)
-	suite.clustering, err = flow.NewClusterList(assignments, suite.nodes)
+	suite.clustering, err = factory.NewClusterList(assignments, suite.nodes)
 	suite.Require().Nil(err)
 
 	suite.epoch.On("Counter").Return(suite.counter, nil)
